Move bytecode compression to `multivm`
