+++ core/lib/multivm/src/versions/vm_fast/tests/bytecode_publishing.rs
@@ -1,7 +1,8 @@
 use zksync_types::event::extract_long_l2_to_l1_messages;
 
 use crate::{
-    interface::{compress_bytecode, TxExecutionMode, VmExecutionMode, VmInterface},
+    interface::{TxExecutionMode, VmExecutionMode, VmInterface},
+    utils::bytecode,
     vm_fast::tests::{
         tester::{DeployContractsTx, TxType, VmTesterBuilder},
         utils::read_test_contract,
@@ -21,7 +22,7 @@ fn test_bytecode_publishing() {
     let counter = read_test_contract();
     let account = &mut vm.rich_accounts[0];
 
-    let compressed_bytecode = compress_bytecode(&counter).unwrap();
+    let compressed_bytecode = bytecode::compress(counter.clone()).unwrap().compressed;
 
     let DeployContractsTx { tx, .. } = account.get_deploy_tx(&counter, None, TxType::L2);
     vm.vm.push_transaction(tx);
