+++ core/lib/multivm/src/versions/vm_latest/tests/bytecode_publishing.rs
@@ -1,7 +1,8 @@
 use zksync_types::event::extract_long_l2_to_l1_messages;
 
 use crate::{
-    interface::{compress_bytecode, TxExecutionMode, VmExecutionMode, VmInterface},
+    interface::{TxExecutionMode, VmExecutionMode, VmInterface},
+    utils::bytecode,
     vm_latest::{
         tests::{
             tester::{DeployContractsTx, TxType, VmTesterBuilder},
@@ -24,7 +25,7 @@ fn test_bytecode_publishing() {
     let counter = read_test_contract();
     let account = &mut vm.rich_accounts[0];
 
-    let compressed_bytecode = compress_bytecode(&counter).unwrap();
+    let compressed_bytecode = bytecode::compress(counter.clone()).unwrap().compressed;
 
     let DeployContractsTx { tx, .. } = account.get_deploy_tx(&counter, None, TxType::L2);
     vm.vm.push_transaction(tx);
